UTF-8 / non-ASCII character support for filenames
In this updated function, the file name is explicitly decoded as UTF-8 using encode('utf-8').decode('unicode_escape'). This should properly handle non-ASCII characters in the file name when printing and saving the file. Using repr() to print the file name will display the escaped version of the file name, making it easier to see the actual characters.